\docType{package}
\name{monographaR}
\alias{monographaR}
\alias{monographaR-package}
\title{Tools for taxonomic monographs.}
\description{
monographaR contains functions intended to facilitate the production
of plant taxonomic monographs. The package includes functions to convert 
tables into taxonomic descriptions, lists of collectors and examined specimens. 
Additionally, wrapper functions to batch the production of phenology histograms
and distributional and diversity maps are also available. 
}


\details{
\tabular{ll}{
Package: \tab monographaR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-08-01\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Marcelo Reginato

Maintainer: Marcelo Reginato <reginatobio@yahoo.com.br>

}
\references{
Reginato, M. (submitted) monographaR: an R package to facilitate the production of plant taxonomic monographs. Brittonia.
}

\keyword{ taxonomy, monographs }
\seealso{
\code{\link[circular:circular-package]{circular}}
\code{\link[maptools:readShapePoly]{maptools}}
\code{\link[raster:raster-package]{raster}}
\code{\link[sp:sp]{sp}}

}

