\name{monographaR_examples}
\alias{monographaR_examples}
\docType{data}
\title{
Input data examples
}
\description{
Input table examples. Seven data.frames are listed in this example data set. 
See help files of the functions for details.
}
\examples{

data(monographaR_examples)

names(monographaR_examples)

head(monographaR_examples$colletorList)
head(monographaR_examples$examinedSpecimens)
head(monographaR_examples$phenoHist)
head(monographaR_examples$tableToDescription)
head(monographaR_examples$map_data)
head(monographaR_examples$mapPhenology)
head(monographaR_examples$taxonomic_headings)


}

