\docType{package}
\name{monographaR}
\alias{monographaR}
\alias{monographaR-package}
\title{Tools for taxonomic monographs.}
\description{
monographaR contains functions intended to facilitate the production
of plant taxonomic monographs. The package includes functions to convert 
tables into taxonomic descriptions, lists of collectors, examined specimens, 
dichotomous keys, and can generate a monograph skeleton. Additionally, 
wrapper functions to batch the production of phenology histograms and 
distributional and diversity maps are also available. 
}


\details{
\tabular{ll}{
Package: \tab monographaR\cr
Type: \tab Package\cr
Version: \tab 1.3.1\cr
Date: \tab 2024-02-01\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Marcelo Reginato

Maintainer: Marcelo Reginato <reginatobio@yahoo.com.br>

}
\references{
Reginato, M. (2016) monographaR: an R package to facilitate the production 
of plant taxonomic monographs. Brittonia 68(2): 212-216.
}

\keyword{ monograph }
\concept{ taxonomy }
\concept{ systematics }
\seealso{
\code{\link[circular:circular-package]{circular}}
\code{\link[raster:raster-package]{raster}}
\code{\link[sp:sp]{sp}}
\code{\link[rmarkdown:rmarkdown-package]{rmarkdown}}
\code{\link[rpart:rpart]{rpart}}

}

